docs: submitting-patches: Add Sponsored-by tag to give credits to who sponsored the patch
Sometimes it happens that a Company or a Physical Person sponsors the
creation and/or the upstreaming process of a patch, but at the moment
there is no way to give credits to it. There are some commit that include
a sort of tag "Sponsored by" without the dash to avoid
scripts/checkpatch.pl to complain but a real standard has not been defined.
With this patch let's try to define a method to give credits consistently
including an acknowledge from the sponsor. The goal is to improve
contributions from companies or physical persons that this way should gain
visibility in Linux kernel and so they should be more prone to let the
work they payed for to be upstreamed.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>